backend/microservices/meal_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.5
requests==2.32.3
pydantic==1.10.14
pillow==10.4.0
python-multipart==0.0.9
